Use average of historical data.

+++ planetdisplay.py
@@ -120,7 +120,7 @@ class PlanetDisplay(object):
                     if self.selected == (xo,y):
                         color = (255, 255, 0)
                     elif h > 0:
-                        value = 255 - 25 * h
+                        value = int(255 - 25 * h)
                         color = (value, value, value)
                     else:
                         color = (255, 255, 255)
